Port membership pages to wrapper
The membership page had some embedded Javascript powering the two
Stripe buttons. I moved that out into a dedicated JS file (under
assets) and added a conditional bit to the wrapper that allows a page
to say "hey, I have a bit of javascript to add!" (just in case there's
any more of this sort of thing I haven't run across yet).

The old site was loading the Stripe JS on every page. I _think_ it's
only needed on this page, to support that custom script. Accordingly,
I moved that <script> tag to be inside the "I have a custom script"
conditional. If/when that needs to get more complicated, the Stripe
bit probably should be controlled by a distinct boolean in the front
matter of the pages that need it.